+++ app/src/main/java/com/example/rodofroll/Objetos/Monstruo.java
@@ -13,20 +13,20 @@ public class Monstruo extends Combatiente {
         this.key=key;
 
         //Atributos
-        vida = Double.parseDouble(this.atributos.get("vida").toString());
-        modiniciativa = Double.parseDouble(this.atributos.get("iniciativa").toString());
-        armadura = Double.parseDouble(this.atributos.get("armadura").toString());
-        ataque = Double.parseDouble(this.atributos.get("ataque").toString());
-        velocidad = Double.parseDouble(this.atributos.get("velocidad").toString());
-        fuerza = Double.parseDouble(this.atributos.get("fuerza").toString());
-        destreza = Double.parseDouble(this.atributos.get("destreza").toString());
-        constitucion = Double.parseDouble(this.atributos.get("constitucion").toString());
-        inteligencia = Double.parseDouble(this.atributos.get("inteligencia").toString());
-        sabiduria = Double.parseDouble(this.atributos.get("sabiduria").toString());
-        carisma = Double.parseDouble(this.atributos.get("carisma").toString());
-        fortaleza = Double.parseDouble(this.atributos.get("fortaleza").toString());
-        reflejos = Double.parseDouble(this.atributos.get("reflejos").toString());
-        voluntad = Double.parseDouble(this.atributos.get("voluntad").toString());
+        vida = Integer.parseInt(this.atributos.get("vida").toString());
+        modiniciativa = Integer.parseInt(this.atributos.get("iniciativa").toString());
+        armadura = Integer.parseInt(this.atributos.get("armadura").toString());
+        ataque = Integer.parseInt(this.atributos.get("ataque").toString());
+        velocidad = Integer.parseInt(this.atributos.get("velocidad").toString());
+        fuerza = Integer.parseInt(this.atributos.get("fuerza").toString());
+        destreza = Integer.parseInt(this.atributos.get("destreza").toString());
+        constitucion = Integer.parseInt(this.atributos.get("constitucion").toString());
+        inteligencia = Integer.parseInt(this.atributos.get("inteligencia").toString());
+        sabiduria = Integer.parseInt(this.atributos.get("sabiduria").toString());
+        carisma = Integer.parseInt(this.atributos.get("carisma").toString());
+        fortaleza = Integer.parseInt(this.atributos.get("fortaleza").toString());
+        reflejos = Integer.parseInt(this.atributos.get("reflejos").toString());
+        voluntad = Integer.parseInt(this.atributos.get("voluntad").toString());
 
         //Biografia
         nombre= (String) this.biografia.get("nombre");
